+++ dev/src/SignalBinding.js
@@ -1,16 +1,20 @@
 	
+	// SignalBinding -------------------------------------------------
+	//================================================================
+	
 	/**
 	 * Object that represents a binding between a Signal and a listener function.
-	 * <br />- <strong>Constructor shouldn't be called by regular user, used internally.</strong>
+	 * <br />- <strong>This is an internall constructor and shouldn't be called by regular user.</strong>
 	 * <br />- inspired by Joa Ebert AS3 SignalBinding and Robert Penner's Slot classes.
 	 * @author Miller Medeiros
 	 * @constructor
+	 * @name signals.SignalBinding
+	 * @param {signals.Signal} signal	Reference to Signal object that listener is currently bound to.
 	 * @param {Function} listener	Handler function bound to the signal.
 	 * @param {boolean} isOnce	If binding should be executed just once.
-	 * @param {?Object} listenerContext	Context on which listener will be executed (object that should represent the `this` variable inside listener function).
-	 * @param {signals.Signal} signal	Reference to Signal object that listener is currently bound to.
+	 * @param {Object} [listenerContext]	Context on which listener will be executed (object that should represent the `this` variable inside listener function).
 	 */
-	signals.SignalBinding = function(listener, isOnce, listenerContext, signal){
+	 function SignalBinding(signal, listener, isOnce, listenerContext){
 		
 		/**
 		 * Handler function bound to the signal.
@@ -28,7 +32,9 @@
 		
 		/**
 		 * Context on which listener will be executed (object that should represent the `this` variable inside listener function).
-		 * @type Object
+		 * @memberOf signals.SignalBinding.prototype
+		 * @name context
+		 * @type {Object|undefined}
 		 */
 		this.context = listenerContext;
 		
@@ -38,10 +44,9 @@
 		 * @private
 		 */
 		this._signal = signal;
-	};
+	}
 	
-	
-	signals.SignalBinding.prototype = {
+	SignalBinding.prototype = /** @lends signals.SignalBinding.prototype */ {
 		
 		/**
 		 * @type boolean
@@ -59,7 +64,9 @@
 			var r;
 			if(this._isEnabled){
 				r = this._listener.apply(this.context, paramsArr);
-				if(this._isOnce) this.detach();
+				if(this._isOnce){
+					this.detach();
+				}
 			}
 			return r; //avoid warnings on some editors
 		},
@@ -82,7 +89,7 @@
 		
 		/**
 		 * Remove binding from signal and destroy any reference to external Objects (destroy SignalBinding object).
-		 * <br /> - calling methods on the binding instance after calling dispose will throw errors.
+		 * <p><strong>IMPORTANT:</strong> calling methods on the binding instance after calling dispose will throw errors.</p>
 		 */
 		dispose : function(){
 			this.detach();