+++ dev/src/intro.js
@@ -1,8 +1,0 @@
-/*!!
- * JS Signals <https://github.com/millermedeiros/js-signals>
- * Released under the MIT license (http://www.opensource.org/licenses/mit-license.php)
- * @author Miller Medeiros <http://millermedeiros.com>
- * @version ::VERSION_NUMBER::
- * @build ::BUILD_NUMBER:: ::BUILD_DATE::
- */
-(function(){