+++ dev/src/signals.js
@@ -1,21 +1,30 @@
-	
+/*jslint onevar:true, undef:true, newcap:true, regexp:true, bitwise:true, maxerr:50, indent:4, white:false, nomen:false, plusplus:false */
+
+/*!!
+ * JS Signals <http://millermedeiros.github.com/js-signals/>
+ * Released under the MIT license <http://www.opensource.org/licenses/mit-license.php>
+ * @author Miller Medeiros <http://millermedeiros.com/>
+ * @version ::VERSION_NUMBER::
+ * @build ::BUILD_NUMBER:: (::BUILD_DATE::)
+ */
+var signals = (function(){
+
 	/**
 	 * @namespace Signals Namespace - Custom event/messaging system based on AS3 Signals
 	 * @name signals
 	 */
-	var signals = window.signals = {};
-	
+	var signals = {};
+
 	/**
 	 * Signals Version Number
 	 * @type string
 	 * @const
 	 */
 	signals.VERSION = '::VERSION_NUMBER::';
+
+//::SIGNAL_BINDING_JS:://
+//::SIGNAL_JS:://
+
+	return signals;
 	
-	/**
-	 * @param {*} param	Parameter to check.
-	 * @return {boolean} `true` if parameter is different than `undefined`.
-	 */
-	signals.isDef = function(param){
-		return typeof param !== 'undefined';
-	};
+}());